+++ src/index.js
@@ -1,169 +1,13 @@
-import velocity from 'velocity-animate';
-
-/**
- * Print an error message
- *
- * @private
- */
-function _error(type, method) {
-  if(type === 1) console.warn(`No elements selected by ${method}`);
-};
-
-
-/**
- * Adds a new selector array or a new selector element to an existing selector array.
- *
- * @private
- * @param {Array} Default array to save
- * @param {Array|Object} Element or array of elements
- * @returns {Array} Element selector array
- */
-function _concat(beforeArr, afterArr) {
-  afterArr = Array.prototype.concat.call([], afterArr);
-  afterArr.forEach(el => {
-    if(beforeArr.indexOf(el) === -1) beforeArr = Array.prototype.concat.call(beforeArr, el);
-  });
-  return beforeArr;
-};
-
-/**
- * Finds the element in the DOM by separating the selector string.
- *
- * @private
- * @param {String} Element selector
- * @returns {Object} Element selector
- */
-function _select(element) {
-  element = element.trim();
-  if(/[\,\>\: \]\=]/.test(element)) {
-    if(!document.querySelectorAll(element).length) element = [];
-    else element = _arrayElement([], document.querySelectorAll(element));
-  } else {
-    switch(element[0]) {
-      case '#' :
-        element = _arrayElement([], document.getElementById(element.substring(1)));
-        break;
-      case '.' :
-        if(!document.getElementsByClassName(element.substring(1)).length) element = [];
-        else element = _arrayElement([], document.getElementsByClassName(element.substring(1)));
-        break;
-      default :
-        if(!document.getElementsByTagName(element).length) element = [];
-        else element = _arrayElement([], document.getElementsByTagName(element));
-    }
-  }
-  return element;
-};
-
-
-/**
- * Put selected element in class.
- *
- * @private
- * @param {String} Element selector
- */
-function _selectElement(element) {
-  if(element === window || element === document) {
-    this.element = _arrayElement([], element);
-  } else if(typeof element === 'object') {
-    this.element = element;
-  }
-
-  if(typeof element === 'string') {
-    this.element = _select(element);
-  }
-};
-
-/**
- * Returns an array of the corresponding elements of the selector
- *
- * @private
- * @param {Array} Default array to save
- * @param {String} Element selector
- * @returns {Array} Element selector array
- */
-function _arrayElement(store, element) {
-  // const select = _select(element);
-  if(!element) return store;
-  if(!element.length) return _concat(store, element);
-  for(const target of element) {
-    store = _concat(store, target);
-  }
-  return store;
-};
-
-/**
- * Find the parent element from the child element.
- *
- * @private
- * @param {Array} Array of parent elements
- * @param {Object} Child element
- * @returns {Array|Null} Returns the parent if there is a parent element, or null if there is no parent
- */
-function _findInParent(parent, children) {
-  let cacheParent = children.parentNode;
-  while(cacheParent !== null) {
-    if(parent.indexOf(cacheParent) !== -1) return cacheParent;
-    cacheParent = cacheParent.parentNode;
-  }
-  return null;
-};
-
-
-/**
- * If the value is a function, it executes the function and returns the return value.
- *
- * @private
- * @param {String|Number|Function} Value or response value
- * @param {Boolean} For Number, the default unit is px
- * @returns {String|Number} Final value
- */
-function _finishValue(value, defaultPx=false) {
-  value = typeof value === 'function' ? value() : value;
-  if(defaultPx) value = typeof value === 'number' ? value + 'px' : value;
-  return value;
-};
-
-/**
- * Returns the first of the selected elements.
- *
- * @private
- * @param {Object} Loaf-dom class
- * @returns {Object} First element selector
- */
- function _oneSelect() {
-  return this.element.length ? this.element[0] : this.element;
-};
-
-/**
- * Divide the specified string into an array, and then exclude the false element.
- *
- * @private
- * @param {String} Strings before division into an array
- * @param {String} String to divide
- * @returns {Array} Compact arrangement
- */
-function _compactSplit(str, value) {
-  return str.split(value).filter(Boolean);
-};
-
-/**
- * Returns an array of the unique values of the two arrays.
- *
- * @private
- * @returns {Array} An array of unique values
- */
-function _union(arr1, arr2) {
-  return [...new Set([...arr1, ...arr2])];
-};
+import Util from './services/util'
+import Err from './services/error'
 
 class LoafDom {
 
   constructor(element) {
-    this.version = '0.3.9';
+    this.version = '0.4.0';
     this.element = [];
-    _selectElement.call(this, element);
-    if(!this.element.length) return _error(1, 'constructor()');
+    Util.selectElement.call(this, element);
+    if(!this.element.length) return Err.warn(1, 'constructor()');
 
     return this;
   }
@@ -177,7 +21,7 @@ class LoafDom {
    */
   el(idx) {
     idx = typeof idx === 'number' ? idx : 0;
-    if(!this.element.length || typeof this.element[idx] === 'undefined') return _error(1, 'el()');
+    if(!this.element.length || typeof this.element[idx] === 'undefined') return Err.warn(1, 'el()');
     return this.element[idx];
   }
   /**
@@ -199,7 +43,7 @@ class LoafDom {
    */
   eq(idx) {
     idx = typeof idx === 'number' ? idx : 0;
-    if(!this.element.length || typeof this.element[idx] === 'undefined') return _error(1, 'eq()');
+    if(!this.element.length || typeof this.element[idx] === 'undefined') return Util.error(1, 'eq()');
     else this.element = [this.element[idx]];
     return this;
   }
@@ -212,7 +56,7 @@ class LoafDom {
    * @returns {Object} New class Loaf-DOM
    */
   addClass(...className) {
-    const el = _oneSelect.call(this);
+    const el = Util.oneSelect.call(this);
     el.classList.add(...className);
     return this;
   }
@@ -239,8 +83,8 @@ class LoafDom {
    * @returns {Boolean} Whether you have the class
    */
   hasClass(className) {
-    const el = _oneSelect.call(this);
-    const baseClassName = _compactSplit(el.className, ' ');
+    const el = Util.oneSelect.call(this);
+    const baseClassName = Util.compactSplit(el.className, ' ');
     return baseClassName.indexOf(className) !== -1;
   }
 
@@ -253,8 +97,8 @@ class LoafDom {
    * @returns {Object|String} New class Loaf-DOM or Attribute Value
    */
   attr(key, value=null) {
-    if(!value) return _oneSelect.call(this).getAttribute(key);
-    _oneSelect.call(this).setAttribute(key, _finishValue(value));
+    if(!value) return Util.oneSelect.call(this).getAttribute(key);
+    Util.oneSelect.call(this).setAttribute(key, Util.finishValue(value));
     return this;
   }
 
@@ -267,8 +111,8 @@ class LoafDom {
    * @returns {Object|String} New class Loaf-DOM or Style Value
    */
   style(key, value=null) {
-    if(!value) return _oneSelect.call(this).style[key];
-    this.element.forEach(el => el.style[key] = _finishValue(value, true));
+    if(!value) return Util.oneSelect.call(this).style[key];
+    this.element.forEach(el => el.style[key] = Util.finishValue(value, true));
     return this;
   }
 
@@ -314,11 +158,11 @@ class LoafDom {
     this.element.forEach(el => {
       for(const child of el.children) {
         if(!selectChild) {
-          store = _concat(store, child);
+          store = Util.concat(store, child);
         } else {
-          const selectChildEl = _arrayElement([], _select(selectChild));
+          const selectChildEl = Util.arrayElement([], Util.select(selectChild));
           if(selectChildEl.indexOf(child) !== -1) {
-            store = _concat(store, child);
+            store = Util.concat(store, child);
           }
         }
       }
@@ -334,39 +178,12 @@ class LoafDom {
    * @returns {Object} New selector dom class
    */
   parents(selectParent) {
-    const selectParentEl = _arrayElement([], _select(selectParent));
+    const selectParentEl = Util.arrayElement([], Util.select(selectParent));
     let store = [];
     this.element.forEach(el => {
-      store = _concat(store, _findInParent(selectParentEl, el));
+      store = Util.concat(store, Util.findInParent(selectParentEl, el));
     });
     return new this.constructor(store);
-  }
-
-  /**
-   * It gives dynamic change.
-   *
-   * @static
-   * @param {Object} Velocity Arguments
-   * @returns {Object} New class Loaf-DOM
-   */
-  animate(...attr) {
-    this.element.forEach(el => {
-      velocity(el, ...attr);
-    });
-    return this;
-  }
-
-  /**
-   * Stops all animation effects on the selected element.
-   *
-   * @static
-   * @returns {Object} New class Loaf-DOM
-   */
-  stop() {
-    this.element.forEach(el => {
-      velocity(el, 'stop');
-    });
-    return this;
   }
 
   /**
@@ -437,7 +254,7 @@ class LoafDom {
    * @returns {Object} offset value
    */
   offset() {
-    const el = _oneSelect.call(this);
+    const el = Util.oneSelect.call(this);
     return { top: el.offsetTop, left: el.offsetLeft, width: el.offsetWidth, height: el.offsetHeight }
   }
 
@@ -449,8 +266,8 @@ class LoafDom {
    * @returns {Number|Object} The width value of the first element | New class Loaf-DOM
    */
   width(widthValue = null) {
-    if(!widthValue) return _oneSelect.call(this).clientWidth;
-    this.style('width', _finishValue(widthValue, true));
+    if(!widthValue) return Util.oneSelect.call(this).clientWidth;
+    this.style('width', Util.finishValue(widthValue, true));
     return this;
   }
 
@@ -462,8 +279,8 @@ class LoafDom {
    * @returns {Number|Object} The height value of the first element | New class Loaf-DOM
    */
   height(heightValue = null) {
-    if(!heightValue) return _oneSelect.call(this).clientHeight;
-    this.style('height', _finishValue(heightValue, true));
+    if(!heightValue) return Util.oneSelect.call(this).clientHeight;
+    this.style('height', Util.finishValue(heightValue, true));
     return this;
   }
 
@@ -475,8 +292,8 @@ class LoafDom {
    * @returns {Number|Object} The scroll position of the top of the element | New class Loaf-DOM
    */
   scrollTop(positionValue = null) {
-    if(!positionValue) return _oneSelect.call(this).scrollTop;
-    _oneSelect.call(this).scrollTop = _finishValue(positionValue);
+    if(!positionValue) return Util.oneSelect.call(this).scrollTop;
+    Util.oneSelect.call(this).scrollTop = Util.finishValue(positionValue);
     return this;
   }
 
@@ -488,8 +305,8 @@ class LoafDom {
    * @returns {Number|Object} The scroll position of the left of the element | New class Loaf-DOM
    */
   scrollLeft(positionValue = null) {
-    if(!positionValue) return _oneSelect.call(this).scrollLeft;
-    _oneSelect.call(this).scrollLeft = _finishValue(positionValue);
+    if(!positionValue) return Util.oneSelect.call(this).scrollLeft;
+    Util.oneSelect.call(this).scrollLeft = Util.finishValue(positionValue);
     return this;
   }
 
@@ -500,7 +317,7 @@ class LoafDom {
    * @returns {Number} The height value of the scroll
    */
   scrollHeight() {
-    return _oneSelect.call(this).scrollHeight;
+    return Util.oneSelect.call(this).scrollHeight;
   }
 
   /**
@@ -510,7 +327,7 @@ class LoafDom {
    * @returns {Number} The width value of the first element
    */
   scrollWidth() {
-    return _oneSelect.call(this).scrollWidth;
+    return Util.oneSelect.call(this).scrollWidth;
   }
 
   /**
@@ -521,9 +338,9 @@ class LoafDom {
    * @returns {String|Object} The html element or New class Loaf-DOM
    */
   html(htmlValue = null) {
-    if(!htmlValue) return _oneSelect.call(this).innerHTML;
+    if(!htmlValue) return Util.oneSelect.call(this).innerHTML;
     this.element.forEach(el => {
-      el.innerHTML = _finishValue(htmlValue);
+      el.innerHTML = Util.finishValue(htmlValue);
     });
     return this;
   }
@@ -536,9 +353,9 @@ class LoafDom {
    * @returns {String|Object} Text of html element or Loaf-dom class
    */
   text(textValue = null) {
-    if(!textValue) return _oneSelect.call(this).innerText;
+    if(!textValue) return Util.oneSelect.call(this).innerText;
     this.element.forEach(el => {
-      el.innerText = _finishValue(textValue);
+      el.innerText = Util.finishValue(textValue);
     });
     return this;
   }
@@ -550,7 +367,7 @@ class LoafDom {
    * @returns {Object} Loaf-dom class
    */
   removeAllChild() {
-    const el = _oneSelect.call(this);
+    const el = Util.oneSelect.call(this);
     while(el.firstChild) {
       el.removeChild(el.firstChild);
     }